src/requirements-old.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --output-file=requirements.txt --pip-args=None requirements.in
#
amqp==5.3.1
    # via
    #   -r requirements.in
    #   kombu
asgiref==3.9.1
    # via
    #   -r requirements.in
    #   django
    #   django-allauth
    #   django-cors-headers
billiard==4.2.2
    # via
    #   -r requirements.in
    #   celery
boto3==1.40.41
    # via -r requirements.in
botocore==1.40.41
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
celery==5.5.3
    # via -r requirements.in
certifi==2025.8.3
    # via
    #   -r requirements.in
    #   requests
charset-normalizer==3.4.3
    # via
    #   -r requirements.in
    #   requests
click==8.3.0
    # via
    #   -r requirements.in
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.1
    # via
    #   -r requirements.in
    #   celery
click-plugins==1.1.1.2
    # via
    #   -r requirements.in
    #   celery
click-repl==0.3.0
    # via
    #   -r requirements.in
    #   celery
dj-rest-auth==7.0.1
    # via -r requirements.in
django==5.2.6
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   django-allauth
    #   django-cors-headers
    #   djangorestframework
    #   djangorestframework-simplejwt
django-allauth==65.11.2
    # via -r requirements.in
django-cors-headers==4.9.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.5.1
    # via -r requirements.in
idna==3.10
    # via
    #   -r requirements.in
    #   requests
jmespath==1.0.1
    # via
    #   -r requirements.in
    #   boto3
    #   botocore
kombu==5.5.4
    # via
    #   -r requirements.in
    #   celery
oauthlib==3.3.1
    # via
    #   -r requirements.in
    #   requests-oauthlib
packaging==25.0
    # via
    #   -r requirements.in
    #   kombu
prompt-toolkit==3.0.52
    # via
    #   -r requirements.in
    #   click-repl
psycopg==3.2.10
    # via -r requirements.in
psycopg-binary==3.2.10
    # via -r requirements.in
pyjwt==2.10.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   botocore
    #   celery
python-dotenv==1.1.1
    # via -r requirements.in
redis==6.4.0
    # via -r requirements.in
requests==2.32.5
    # via
    #   -r requirements.in
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via -r requirements.in
s3transfer==0.14.0
    # via
    #   -r requirements.in
    #   boto3
six==1.17.0
    # via
    #   -r requirements.in
    #   python-dateutil
sqlparse==0.5.3
    # via
    #   -r requirements.in
    #   django
typing-extensions==4.15.0
    # via
    #   -r requirements.in
    #   psycopg
tzdata==2025.2
    # via
    #   -r requirements.in
    #   kombu
urllib3==2.5.0
    # via
    #   -r requirements.in
    #   botocore
    #   requests
vine==5.1.0
    # via
    #   -r requirements.in
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via
    #   -r requirements.in
    #   prompt-toolkit
